# Add span overloads to API