--- a/src/LightningDB.Tests/DatabaseTests.cs
+++ b/src/LightningDB.Tests/DatabaseTests.cs
@@ -89,7 +89,7 @@ namespace LightningDB.Tests
                 using (var cursor = tx.CreateCursor(db))
                 {
                     cursor.MoveNext();
-                    Assert.Equal("customdb", UTF8.GetString(cursor.Current.Key.Span.ToArray()));
+                    Assert.Equal("customdb", UTF8.GetString(cursor.Current.Key.CopyToNewArray()));
                 }
             }
         }