This PR closes #116. It surfaces Span overloads for most operations and should improve both read and write performance primarily  by reducing GC allocations