@AlgorithmsAreCool 2020-06-28
Before : 
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.836 (1909/November2018Update/19H2)
Intel Core i7-8750H CPU 2.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.100-preview.5.20279.10
  [Host]     : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT


```
| Method | OpsPerTransaction | ValueSize |   KeyOrder |         Mean |       Error |      StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |------------------ |---------- |----------- |-------------:|------------:|------------:|--------:|------:|------:|----------:|
|   **Read** |                 **1** |         **8** | **Sequential** |     **424.5 ns** |     **8.15 ns** |     **7.62 ns** |  **0.0710** |     **-** |     **-** |     **336 B** |
|   **Read** |                 **1** |        **64** | **Sequential** |     **429.0 ns** |     **8.51 ns** |    **13.00 ns** |  **0.0830** |     **-** |     **-** |     **392 B** |
|   **Read** |                 **1** |       **256** | **Sequential** |     **441.4 ns** |     **7.69 ns** |     **7.19 ns** |  **0.1240** |     **-** |     **-** |     **584 B** |
|   **Read** |               **100** |         **8** | **Sequential** |  **12,507.1 ns** |   **244.10 ns** |   **372.77 ns** |  **0.7324** |     **-** |     **-** |    **3504 B** |
|   **Read** |               **100** |        **64** | **Sequential** |  **15,080.6 ns** |   **288.58 ns** |   **364.96 ns** |  **1.9226** |     **-** |     **-** |    **9104 B** |
|   **Read** |               **100** |       **256** | **Sequential** |  **16,760.2 ns** |   **331.27 ns** |   **515.74 ns** |  **6.0120** |     **-** |     **-** |   **28304 B** |
|   **Read** |              **1000** |         **8** | **Sequential** | **173,016.9 ns** | **3,397.00 ns** | **8,332.91 ns** |  **6.8359** |     **-** |     **-** |   **32304 B** |
|   **Read** |              **1000** |        **64** | **Sequential** | **179,190.1 ns** | **3,554.55 ns** | **5,428.18 ns** | **18.5547** |     **-** |     **-** |   **88304 B** |
|   **Read** |              **1000** |       **256** | **Sequential** | **201,845.5 ns** | **3,963.95 ns** | **6,512.88 ns** | **59.5703** |     **-** |     **-** |  **280304 B** |

After

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.836 (1909/November2018Update/19H2)
Intel Core i7-8750H CPU 2.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.100-preview.5.20279.10
  [Host]     : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT


```
| Method | OpsPerTransaction | ValueSize |   KeyOrder |         Mean |       Error |       StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |------------------ |---------- |----------- |-------------:|------------:|-------------:|--------:|------:|------:|----------:|
|   **Read** |                 **1** |         **8** | **Sequential** |     **426.7 ns** |     **8.46 ns** |     **11.00 ns** |  **0.0710** |     **-** |     **-** |     **336 B** |
|   **Read** |                 **1** |        **64** | **Sequential** |     **437.0 ns** |     **8.76 ns** |     **15.79 ns** |  **0.0830** |     **-** |     **-** |     **392 B** |
|   **Read** |                 **1** |       **256** | **Sequential** |     **445.2 ns** |     **8.53 ns** |     **10.79 ns** |  **0.1240** |     **-** |     **-** |     **584 B** |
|   **Read** |               **100** |         **8** | **Sequential** |  **12,377.9 ns** |   **236.80 ns** |    **324.14 ns** |  **0.7324** |     **-** |     **-** |    **3504 B** |
|   **Read** |               **100** |        **64** | **Sequential** |  **14,276.2 ns** |   **281.75 ns** |    **438.66 ns** |  **1.9226** |     **-** |     **-** |    **9104 B** |
|   **Read** |               **100** |       **256** | **Sequential** |  **16,474.7 ns** |   **324.40 ns** |    **485.55 ns** |  **6.0120** |     **-** |     **-** |   **28304 B** |
|   **Read** |              **1000** |         **8** | **Sequential** | **165,299.1 ns** | **3,158.93 ns** |  **3,243.99 ns** |  **6.8359** |     **-** |     **-** |   **32304 B** |
|   **Read** |              **1000** |        **64** | **Sequential** | **179,787.0 ns** | **3,389.49 ns** |  **3,170.53 ns** | **18.5547** |     **-** |     **-** |   **88305 B** |
|   **Read** |              **1000** |       **256** | **Sequential** | **220,032.9 ns** | **5,844.55 ns** | **17,048.84 ns** | **59.5703** |     **-** |     **-** |  **280304 B** |




After (with new non allocating API)

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.836 (1909/November2018Update/19H2)
Intel Core i7-8750H CPU 2.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.100-preview.5.20279.10
  [Host]     : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT


```
| Method | OpsPerTransaction | ValueSize |   KeyOrder |         Mean |       Error |      StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |------------------ |---------- |----------- |-------------:|------------:|------------:|-------:|------:|------:|----------:|
|   **Read** |                 **1** |         **8** | **Sequential** |     **449.6 ns** |     **8.88 ns** |    **14.83 ns** | **0.0644** |     **-** |     **-** |     **304 B** |
|   **Read** |                 **1** |        **64** | **Sequential** |     **435.0 ns** |     **8.72 ns** |    **21.88 ns** | **0.0644** |     **-** |     **-** |     **304 B** |
|   **Read** |                 **1** |       **256** | **Sequential** |     **415.3 ns** |     **7.60 ns** |     **8.44 ns** | **0.0644** |     **-** |     **-** |     **304 B** |
|   **Read** |               **100** |         **8** | **Sequential** |  **12,363.8 ns** |   **240.95 ns** |   **337.78 ns** | **0.0610** |     **-** |     **-** |     **304 B** |
|   **Read** |               **100** |        **64** | **Sequential** |  **13,812.4 ns** |   **272.65 ns** |   **505.38 ns** | **0.0610** |     **-** |     **-** |     **304 B** |
|   **Read** |               **100** |       **256** | **Sequential** |  **14,835.0 ns** |   **293.48 ns** |   **521.66 ns** | **0.0610** |     **-** |     **-** |     **304 B** |
|   **Read** |              **1000** |         **8** | **Sequential** | **164,105.6 ns** | **3,256.16 ns** | **5,529.21 ns** |      **-** |     **-** |     **-** |     **304 B** |
|   **Read** |              **1000** |        **64** | **Sequential** | **172,565.8 ns** | **3,365.86 ns** | **5,715.49 ns** |      **-** |     **-** |     **-** |     **304 B** |
|   **Read** |              **1000** |       **256** | **Sequential** | **173,051.3 ns** | **3,410.59 ns** | **7,557.63 ns** |      **-** |     **-** |     **-** |     **304 B** |