@AlgorithmsAreCool 2020-06-28
I need to do another pass over the documentation and do the write benchmarks

I'll take care of those tomorrow, it's late